requirements.txt
Django>=4.2,<5
djangorestframework
django-cors-headers
drf-spectacular
python-decouple
psycopg[binary]>=3.1.0
Pillow
django-mdeditor
django-filter
django-ratelimit
gunicorn
django-environ